Move new grasp loading impl
