+++ molmo_spaces/utils/grasp_sample.py
@@ -10,11 +10,7 @@ import numpy as np
 from scipy.spatial.transform import Rotation as R
 
 from molmo_spaces.env.data_views import MlSpacesObject, create_mlspaces_body
-from molmo_spaces.molmo_spaces_constants import (
-    ASSETS_DIR,
-    OBJECT_LIBRARY_TO_GRASP_LIBRARIES,
-    USER_GRASP_LIBRARIES,
-)
+from molmo_spaces.molmo_spaces_constants import ASSETS_DIR
 from molmo_spaces.robots.robot_views.abstract import RobotView
 from molmo_spaces.utils.asset_names import get_thor_name
 from molmo_spaces.utils.lazy_loading_utils import locate_uid_package, get_user_grasp_library_index
@@ -22,74 +18,6 @@ from molmo_spaces.utils.pose import pos_quat_to_pose_mat
 from molmo_spaces.utils.profiler_utils import Timer
 
 log = logging.getLogger(__name__)
-
-
-def get_grasp_libraries_for_object(uid: str) -> list[str]:
-    package, _, _ = locate_uid_package(uid)
-    if package not in OBJECT_LIBRARY_TO_GRASP_LIBRARIES:
-        return []
-    return list(OBJECT_LIBRARY_TO_GRASP_LIBRARIES[package])
-
-
-def _filter_grasp_libraries_for_object(
-    uid: str, grasp_libraries: list[str] | None = None
-) -> list[str]:
-    available_grasp_libraries = get_grasp_libraries_for_object(uid)
-    if grasp_libraries is None:
-        grasp_libraries = available_grasp_libraries
-    else:
-        available_grasp_libraries = set(available_grasp_libraries)
-        grasp_libraries = [
-            library for library in grasp_libraries if library in available_grasp_libraries
-        ]
-
-    return grasp_libraries
-
-
-def get_grasp_path(uid: str, grasp_libraries: list[str] | None = None) -> Path | None:
-    grasp_libraries = _filter_grasp_libraries_for_object(uid, grasp_libraries)
-
-    for library in grasp_libraries:
-        if library in USER_GRASP_LIBRARIES:
-            grasp_library_dir = USER_GRASP_LIBRARIES[library]
-            grasp_library_index = get_user_grasp_library_index(grasp_library_dir)
-            robot_name = library.split("/", 1)[-1]
-            grasp_file = grasp_library_index.grasp_paths.get(robot_name, {}).get(uid, None)
-            if grasp_file is not None:
-                grasp_file = grasp_library_dir / grasp_file
-        else:
-            grasp_file = ASSETS_DIR / f"grasps/{library}/{uid}/{uid}_grasps_filtered.npz"
-
-        if grasp_file is not None and grasp_file.exists():
-            return grasp_file
-
-    return None
-
-
-def get_joint_grasp_path(
-    uid: str, joint_name: str, grasp_libraries: list[str] | None = None
-) -> Path | None:
-    grasp_libraries = _filter_grasp_libraries_for_object(uid, grasp_libraries)
-
-    for library in grasp_libraries:
-        if library in USER_GRASP_LIBRARIES:
-            grasp_library_dir = USER_GRASP_LIBRARIES[library]
-            grasp_library_index = get_user_grasp_library_index(grasp_library_dir)
-            robot_name = library.split("/", 1)[-1]
-            grasp_file = (
-                grasp_library_index.articulated_grasp_paths.get(robot_name, {})
-                .get(uid, {})
-                .get(joint_name, None)
-            )
-            if grasp_file is not None:
-                grasp_file = grasp_library_dir / grasp_file
-        else:
-            grasp_file = ASSETS_DIR / f"grasps/droid/{uid}/{joint_name}_grasps_filtered.npz"
-
-        if grasp_file is not None and grasp_file.exists():
-            return grasp_file
-
-    return None
 
 
 @functools.lru_cache(
@@ -179,40 +107,6 @@ def has_joint_grasp_file(object_name: str, joint_name: str) -> bool:
         ):  # at least 100 grasps. NOTE(yejin):Attempt to filter out inside drawers like in Fridges
             return True
     return False
-
-
-def load_object_grasps(
-    uid: str, grasp_libraries: list[str] | None = None, num_grasps: int = 50
-) -> np.ndarray:
-    grasp_path = get_grasp_path(uid, grasp_libraries)
-    if grasp_path is None:
-        raise ValueError(f"No grasp file found for {uid}")
-
-    # TODO: wire up this method (and below) into demonstrators
-
-    npz_data = np.load(grasp_path)
-    transforms: np.ndarray = npz_data["transforms"]
-    if len(transforms) <= num_grasps:
-        return transforms
-    else:
-        idxs = random.sample(range(len(transforms)), num_grasps)
-        return transforms[idxs]
-
-
-def load_object_joint_grasps(
-    uid: str, joint_name: str, grasp_libraries: list[str] | None = None, num_grasps: int = 50
-) -> np.ndarray:
-    grasp_path = get_joint_grasp_path(uid, joint_name, grasp_libraries)
-    if grasp_path is None:
-        raise ValueError(f"No joint grasp file found for {uid}/{joint_name}")
-
-    npz_data = np.load(grasp_path)
-    transforms: np.ndarray = npz_data["transforms"]
-    if len(transforms) <= num_grasps:
-        return transforms
-    else:
-        idxs = random.sample(range(len(transforms)), num_grasps)
-        return transforms[idxs]
 
 
 def load_grasps_for_object(object_name, num_grasps=50):
